pyproject.toml
[tool.poetry]
name = "library-agent"
version = "0.1.0"
description = "A Python project using Haystack for document processing and search"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "rag", from = "src"},
    {include = "data", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
python-dotenv = "^1.1.1"
pandas = "^2.3.2"
neo4j = "^5.28.2"
langgraph = "^0.6.7"
langchain-openai = "^0.3.33"
langchain-community = "^0.3.29"
streamlit = "^1.49.1"
langchain-neo4j = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true


